moving retrieval functionality out of generic autocomplete and into specific component for query type
